Acquire spectra interactively

Functions providing a simple interactive front-end to the functions in the package. Mostly as example code that can be modified for diferent uses.

acq_irrad_interactive(tot.time.range = c(5, 15), target.margin = 0.1,
  HDR.mult = c(short = 1, long = 10), protocols = NULL,
  correction.method = NA, descriptors = NA)

acq_fraction_interactive(tot.time.range = c(5, 15), target.margin = 0.2,
  HDR.mult = c(short = 1, long = 10), protocols = NULL,
  correction.method = NA, descriptors = NA, ref.value = 1,
  qty.out = "Tfr", type = "total")

Arguments

tot.time.range
numeric vector Range of total times for a measurement in seconds.

target.margin
numeric (0..1) when tuning integration time, how big a head space to leave.

HDR.mult
numeric the integration time for each bracketed integration as a multiplier of the set or tuned integration time.

protocols
list of character vectors.

correction.method
list The method to use when applying the calibration

descriptors
list A list of instrument descriptors containing calibration data.

ref.value
numeric or filter_spct/reflector_spct object.

qty.out
character One of "Tfr" (transmittance as a fraction of one) or "raw" (raw counts).

type
character Type of transmittance or reflectance measured.

Note

The integration time is set automatically so that the peak number of counts is close to 1 - target.margin of maximum range of the instrument detector. The minmum tot.time is achieved by increasing the number of scans. The default protocols are usually suitable, if new protocols are passed, each character vector must contain strings "light", "filter" and "dark", or "sample", "reference", and "dark", depending on the function.

The calculations for reflectance and transmittance are very similar, so we provide a single function capable of handling both. For transmittance the reference is usually direct exposure to radiation but for reflectance a white reference patch is normally used. In some cases one may want to use a grey reference. We provide an argument that allows the user to supply a constant or a spectrum describing the properties of the reference. It is also important to distinguish between total and internal transmittance, and which of these is measured depends on the measuring protocol.

Examples

not_run({
  # requires an Ocean Optics spectrometer to be connected via USB
  acquire_interactive()
})