Take one set of spectral readings

Take readings according to parameters from a list of settings and a protocol defined by a vector of names.

acq_raw_mspct(descriptor, acq.settings, seq.settings = NULL,
  protocol = c("light", "filter", "dark"), user.label = "",
  where.measured = data.frame(lon = NA_real_, lat = NA_real_),
  pause.fun = NULL, verbose = TRUE, ...)

Arguments

descriptor
list as returned by function get_oo_descriptor()

acq.settings
list as returned by functions tune_acq_settings() or retune_acq_settings() or acq_settings()

protocol
vector of character strings

user.label
character string to set as label

where.measured
data.frame with at least columns "lon" and "lat" compatible with value returned by ggmap::geocode()

pause.fun
function used for handling protocol transitions#'

verbose
ogical to enable or disable warnings

...
passed to pause.fun (ignored by the default function)

Value

a raw_mspct object

Note

The default pause.fun prompts the user at each change of value in protocol. The user can write other functions, for example for a time delay.