Settings for spectral measurement

Validate parameters for spectral measurements and return a list of values usable as input for functions retune_acq_settings(), acq_sptc(), and acq_mspct().

Find optimal settings for spectral measurements under a given measurement protocol.

acq_settings(descriptor, integ.time = 0.01, num.scans = 10L,
  min.integ.time = -Inf, max.integ.time = Inf, tot.time.range = c(0, Inf),
  HDR.mult = c(short = 1, long = 10), target.margin = 0.1,
  pix.selector = TRUE, corr.elect.dark = 0L, corr.sensor.nl = 0L,
  boxcar.width = 0L, verbose = TRUE)

tune_acq_settings(descriptor, acq.settings, verbose = TRUE)

Arguments

descriptor
list as returned by function get_oo_descriptor

integ.time
numeric vaue in seconds

num.scans
integer

min.integ.time
numeric vaue in seconds

max.integ.time
numeric vaue in seconds

tot.time.range
numeric vector of length two with values in seconds

HDR.mult
a numeric vector with integ.time multipliers to be used for "bracketing".

target.margin
numeric (0..1)

pix.selector
a logical or numeric vector used as subscript to select pixels

corr.elect.dark, corr.sensor.nl
integer 0L (FALSE) or 1L (TRUE)

boxcar.width
integer number of pixels to average

verbose
a logical to enable or disable warnings

acq.settings
list as returned by a previous call to acq_settings(), or tune_acq_settings().

Value

a list.

a list.

Details

acq_settings() is used to manually save a complete set of instrument settings in a way that they can be reused as needed for repated acquisition of spectra. tune_acq_settings() and tune_acq_settings() differ only in their formal parameters and optimize settings to the current spectral irradiance.

Note

pixel.selector can be used for two different purposes: to ignore bad pixels and to restrict integration-time tuning to the response from a range of pixels. The interpretation of tot.time.range is as follows: first value is minimum time, second value is maximum time. If both values are the same, then an exact measurement time is computed.

Ocean Optics spectrometers can be queried for the maximum and minimum supported integration times. This function modifies the user supplied values if outside these bounds. The defaults of -Inf and Inf force the use of the whole valid range of integration time supported by the connected intrument. pixel.selector can be used for two different purposes: to ignore bad pixels and to restrict integration-time tuning to the response from a range of pixels.