Correct for stray light

Correct cps readings for stray light, using either masured stray light, or using a non-excited region of the detector array.

filter_correction(x, flt, stray.light.method = "original",
  stray.light.wl = c(218.5, 228.5), flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5), flt.Tfr = 1, trim = 0.05,
  verbose = getOption("photobiology.verbose", default = FALSE))

no_filter_correction(x, stray.light.wl = c(218.5, 228.5),
  flt.dark.wl = c(193, 209.5), flt.ref.wl = NULL, flt.Tfr = 1, trim = 0,
  verbose = getOption("photobiology.verbose", default = FALSE))

Arguments

x, flt
cps_spct objects, containig spectral data from which to subtract stray light, and measured stray light, respectively.

stray.light.method
Method variant used, "original" (Ylianttila), "simple", "full", "sun", "raw", "none".

stray.light.wl
numeric vector of length 2 giving the range of wavelengths to use for the final stray light correction.

flt.dark.wl, flt.ref.wl
numeric vectors of length 2 giving the ranges of wavelengths to use for the "dark" and "illuminated" regions of the array in the filter correction.

flt.Tfr
numeric fractional transmittance of the filter to the source of stray light, used only for method "simple".

trim
a numeric value to be used as argument for mean

verbose
Logical indicating the level of warnings wanted.