Make default method from descriptor

A function that builds a default method from an instrument descriptor. Useful when the spectrometers has not been characterized as needed for the more sofisticated methods. Can use stray light correction but not slit function correction. Stray light correction is valid only if it was also used during irradiance callibration. Suitablity of wavelengths and method depends on the instrument configuration.

new_correction_method(descriptor, stray.light.method = "none",
  stray.light.wl = c(220, 230), flt.dark.wl = c(200, 210),
  flt.ref.wl = c(360, 379.5), flt.Tfr = 1)

Arguments

descriptor
list, as returned by get_oo_descriptr()

stray.light.method
character Name of method.

stray.light.wl, flt.dark.wl, flt.ref.wl
numeric vector with wavelengths (nm).

flt.Tfr
numeric Transmittance of filter as a fraction (0...1).

Value

a list

Details

The currently recognized methods for stray-light correction are "none", "original", "sun", and "simple". With the default method "none", the values of the remaining parameters are ignored.

Note

Defaults for indexes are for the first channel of the first spectrometer currently connected.