Find range of lines in file containing spectral data.

Ocean Optics files can have varying numbers of rows containing spectral data, but these rows are fenced between two lines containing recognizable character strings. This function, detects these strings.

oofile_data_rows(lines)

Arguments

lines
character The data file read line by line.

Value

integer vector of length two, containing the indexes to the first last rows containing spectral data.

Note

This is a rather slow operation, so if all files to be read have the same format, it is inefficient to call this function for each file.