Read File Saved by Ocean Optics' OceanView.

Reads and parses the header of a processed data file as output by OceanView to extract the whole header remark field. The time field is retrieved and decoded.

read_oo_ovdata(file, time = NULL, geocode = NULL, label = NULL,
  descriptor = NULL, tz = NULL, locale = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE))

Arguments

file
character string

time
a POSIXct object, but if NULL the date stored in file is used, and if NA no date variable is added

geocode
A data frame with columns lon and lat.

label
character string, but if NULL the value of file is used, and if NA the "what.measured" attribute is not set.

descriptor
list A list describing the instrument used.

tz
character Time zone is by default read from the file.

locale
The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

verbose
Logical indicating the level of warnings wanted.

Value

A raw_spct object.