Convert raw counts data into spectral irradiance

Convert raw counts data into spectral irradiance

s_irrad_corrected(x, ...)

# S3 method for default
s_irrad_corrected(x, ...)

# S3 method for list
s_irrad_corrected(x, time = NULL, correction.method,
  return.cps = FALSE, descriptor, locale = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE), ...)

# S3 method for raw_mspct
s_irrad_corrected(x, spct.names = c(light = "light",
  filter = "filter", dark = "dark"), correction.method, return.cps = FALSE,
  verbose = getOption("photobiology.verbose", default = FALSE), ...)

# S3 method for raw_spct
s_irrad_corrected(x, correction.method, return.cps = FALSE,
  verbose = getOption("photobiology.verbose", default = FALSE), ...)

Arguments

x
A named list of one to three vectors of file names, with names "light", "filter", and "dark". Or a raw_mspt object, or a raw_spct object.

...
Named argument passed to photobiology::cps2irrad which is the final calculation step.

time
a POSIXct object, but if NULL the date stored in file is used, and if NA no date variable is added

correction.method
A named list of constants and functions defining the method to be sued for stray light and dark signal corrections.

return.cps
logical Useful when there is no need to apply a calibration, such as when computing new calibration multipliers.

descriptor
A named list with a descriptor of the characteristics of the spectrometer (if serial number does not agree an error is triggered).

locale
The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

verbose
Logical indicating the level of warnings wanted.

spct.names
named character vector of length three, to map names in x to those expected.

Methods (by class)

default: Default for generic function.

list: Default for generic function.

raw_mspct: Default for generic function.

raw_spct: Default for generic function.

Note

Currently s_irrad_corrected.list allows processing of files written by OceanOptics' SpectraSuite software, from protocols with integration-time bracketing or not, with a dark reference measurement or not. Three measurements components are recognized: a "light" measurement, a "filter" measurement using a polycarbonate filter and a dark measurement. Only "light" is mandatory. Data should be raw counts, either corrected for detector non-linearity or not. All three spectra should be acquired using the same instrument settings to achieve good accuracy.