Apply filter-based stray-light correction

Apply to a counts-per-second spectrum corrections based of a paired reading obtained with a polycarbonate filter (long-pass with cut-off at 400 nm). This is a bit more sofisticated than simple subtracting the filter reading from the mesurement as the effect of the filter itself on stray light is corrected for.

uvb_corrections(x, spct.names = c(light = "light", filter = "filter", dark =
  "dark"), stray.light.method = "original", stray.light.wl = c(218.5,
  228.5), flt.dark.wl = c(193, 209.5), flt.ref.wl = c(360, 379.5),
  flt.Tfr = 0.9, inst.dark.pixs = 1:4, worker.fun = NULL, trim = 0.05,
  verbose = getOption("photobiology.verbose", default = FALSE), ...)

slit_function_correction(x, worker.fun = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE), ...)

Arguments

x
raw_mspct The raw counts from measurements.

spct.names
named character vector of length three.

stray.light.method
Method variant used, "original" (Ylianttila), "simple", "full", "sun", "raw", "none".

stray.light.wl
numeric vector of length 2 giving the range of wavelengths to use for the final stray light correction.

flt.dark.wl, flt.ref.wl
numeric vectors of length 2 giving the ranges of wavelengths to use for the "dark" and "illuminated" regions of the array in the filter correction.

flt.Tfr
numeric fractional transmittance of the filter to the source of stray light, used only for method "simple".

inst.dark.pixs
numeric vector with indexes to array pixels that are in full drakness by instrument design.

worker.fun
function actually doing the correction on the w.lengths and counts per second vectors.

trim
a numeric value to be used as argument for mean

verbose
Logical indicating the level of warnings wanted.

...
additional params passed to worker.fun.

Value

A cps_spct object with the corrected count-per-second values in coulmn "cps".

Note

stray.light.method = "none" is a valid argument only for function uvb_corrections(). The default worker.fun is just an example. Corrections are specific to each individual spectrometer unit (not just type or configuration) and code needs to be written for most individual use cases. The slit function tail correction requires the characterization of the shape of the slit function by measuring one or more laser beams at suitable wavelengths.

References

http://www.r4photobiology.info