Select which instrument descriptor to use

Select from a list of instrument descriptors which one to use based on date of measurement.

which_descriptor(date = lubridate::today(),
  descriptors = ooacquire::MAYP11278_descriptors, verbose = NULL, ...)

Arguments

date
Any object that anytime::anydate() will decode as a date or convert to a date. Used to select a descriptor containing calibration data valid for a given day.

descriptors
A named list of descriptors of the characteristics of the spectrometer including calibration data.

verbose
Logical indicating the level of warnings wanted.

...
Currently ignored.

Note

The default argument for verbose is for this function TRUE unless the argument to date is a date object, as conversion of other objects to a date may fail.

If a character string is passed as argument to date, it must be in a format suitable for anytime::anydate(). One needs to be careful with months and days of the month when supplying them as numbers, so using months names or their abbreviations may be safer.